[PATCH] fbdev: Add mode changing via sysfs
This is a first pass at adding two new sysfs attributes to
/sys/class/graphics/fb0 for setting modes.  There are two attributes: modes
which contains a list of valid modes, and mode which is the current mode.  To
switch modes echo one of the entries from the modes list to the mode
attribute.

The D,V,S on the modes represents Detailed, Vesa, Standard from the DDC info.

modes is root writable.  It can also be used to set the list of modes.  For
example a /etc file could add modes that are not in the monitor's DDC.

mode is user writable.  PAM would set ownership of mode at user login time.
This provides a safe way for a user to set the mode without being root.  You
can only set the mode to one of the modes on the list.

Signed-off-by: Antonino Daplas <adaplas@pol.net>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 42307fefql2TKmDj6Wi1Cv_7aiv0ng