[PATCH] dvb: add pll lib
This adds some helper code to handle tuning for dvb cards, with a struct
describing the pll and a function calculating the command sequence needed to
program it.

This one was discussed + accepted on the linuxtv list and also is in the
linuxtv cvs.  As the the cx88 driver update I want finally get out of the door
depends on this one I'll go submit it myself instead of waiting for the dvb
guys doing it.

Signed-off-by: Gerd Knorr <kraxel@bytesex.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 423319a36q-_c0uFUy7yn7K9DzXlzw