Merge bk://gkernel.bkbits.net/net-drivers-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2005/03/11 23:27:01-05:00 alan
[PATCH] -ac bits for ULI ethernet missing from 2.6.11

Signed-off-by: Jeff Garzik <jgarzik@pobox.com>

2005/03/11 23:26:46-05:00 willy
[PATCH] Lance needs delay.h

An 'allyesconfig' on PA-RISC shows that lance is using mdelay without
including <linux/delay.h>

Signed-off-by: Matthew Wilcox <matthew@wil.cx>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>

2005/03/11 23:26:34-05:00 apatard
[PATCH] sk98lin driver: fix driver name string

I've just notice that in 2.6.11 in the sk98lin driver, the name of the
module is set to skge in the pci_driver structure. IMHO, it should be
sk98lin.

Signed-off-by: Jeff Garzik <jgarzik@pobox.com>

2005/03/11 23:26:21-05:00 romieu
[PATCH] Fix r8169: panic on 2.6.11

Stephen Hemminger <shemminger@osdl.org> :
[...]
> > +	u32 eor = le32_to_cpu(desc->opts1) & ~RingEnd;
                                             ^
Oops.

Updated patch against plain 2.6.11 below (NAPI is strongly
suggested, no need to disable rx_copybreak). The patch has
been tested on a 100Mbps ethernet but it does not want to
crash. I'll hack the testbed and see tomorrow how things
behave on Gb.

The status and received packets indication in the Rx descriptor ring
are not correctly reset when a descriptor is recycled.

Signed-off-by: Francois Romieu <romieu@fr.zoreil.com>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>

2005/03/11 23:26:08-05:00 bunk
[PATCH] drivers/net/typhoon: make a firmware image static

This patch makes a nedlessly global firmware image static.

This patch was already ACK'ed by David Dillow.

Signed-off-by: Adrian Bunk <bunk@stusta.de>
Signed-off-by: Jeff Garzik <jgarzik@pobox.com>

BKrev: 423322a3FwKW9HzI5LcF6gKCVV86wA