[PATCH] PA-RISC GSC interrupt fixes
Two problems occurred with the recent switch to virtual interrupt numbers.

The first is that two devices (eg Lasi PS/2 mouse and keyboard) can
share the same physical interrupt and should therefore share the same
virtual interrupt rather than have a new one assigned to them.  Fix this
by initialising the global_irq array to NO_IRQ, then checking the array
before assigning a new one.

The second is that some GSC devices (eg HIL and EISA) can have
two interrupts; one for normal interrupts and the other for NMI or
high-priority.  To fix this, I introduced an auxiliary IRQ to the
parisc_device and fill it in in Asp and Wax.

Signed-off-by: Matthew Wilcox <willy@parisc-linux.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4233238bvgRBWMf8-RLOJIXCgEW6iw