[PATCH] x86: fix booting non-NUMA system with NUMA config
There is a bug booting with CONFIG_NUMA=y, CONFIG_X86_GENERICARCH=y, and
booting on a non-NUMA system.  While not the most common configuration, it
should surely be supported.

memmap_init_zone() is the first user to do pfn_to_nid(), which relies on
physnode_map[] to be done properly.  memory_present() was supposed to do
that, but never got called for the flat configuration, so pfn_to_nid() was
returning -1 on valid pages.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4234d7adpwynckE9qdHAdknUISF8EA