[PATCH] swsusp: enable resume from initrd
From: <mjg59@scrf.ucam.org>

When using a fully modularized kernel it is necessary to activate resume
manually as the device node might not be available during kernel init.

This patch implements a new sysfs attribute '/sys/power/resume' which allows
for manual activation of software resume.  When read from it prints the
configured resume device in 'major:minor' format.  When written to it expects
a device in 'major:minor' format.  This device is then checked for a suspended
image and resume is started if a valid image is found.  The original
functionality is left in place.

It should be used from initramfs, or with care.

Signed-off-by: Hannes Reinecke <hare@suse.de>
Signed-off-by: Pavel Machek <pavel@suse.cz>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 4234d9ebK-zP3XDhy3L62U55_bXWXw