Merge bk://kernel.bkbits.net/davem/net-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2005/03/14 21:43:50-08:00 Robert.Olsson
[PKTGEN]: Fix thread_lock() usage.

The current code bombs out with preemption enabled.

Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:41:19-08:00 herbert
[NETFILTER]: ipt_hashlimit.c needs lockhelp.h

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:39:50-08:00 shemminger
[BRIDGE]: No update when hold time is zero

Some users, set hold time to zero on bridge so it always does
flooding. This is usually when using it with wireless.  The new RCU
based code changed the behaviour so the bridge would not flood for
one GC interval. This patch restores the original behaviour.

Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:38:52-08:00 herbert
[IPSEC]: Do not hold state lock while checking size.

This can elicit ICMP message output and thus result in a
deadlock.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:36:52-08:00 andre
[TCP]: Fix gcc-3.4 build when TCP_DEBUG is disabled in tcp.h

Signed-off-by: Andre Tomt <andre@tomt.net>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:35:35-08:00 laforge
[NETFILTER]: ipt_hashlimit: use | instead of & for correct hash key

Signed-off-by: Harald Welte <laforge@netfilter.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:29:54-08:00 ralf
[AX25]: Fix ax25_get_socket locking

In an attempt to return a locked socket ax25_get_socket() was calling
lock_sock() with a spinlock held, bad idea.  Making matters worse it's
only user is running in bottom half context resulting in a potencial
attempt to sleep in bottom half context, so fix the locking there as well.

Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:28:10-08:00 herbert
[NETLINK]: Fix multicast bind/autobind race

netlink_autobind has always set nlk_sk(sk)->groups to zero.  This is
unnecessary because sk_alloc already zeroes the entire structure.
Since a socket can only be bound once netlink_autobind doesn't need
to zero groups at all.

This had been safe until I added mc_list.  Now it is possible for
netlink_bind to race against netlink_autobind running on the same
socket on another CPU.  The result would be a socket that's on
mc_list with groups set to zero.  This socket will be left on the
list even after it is destroyed.

The fix is to remove the zeroing in netlink_autobind.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:27:07-08:00 herbert
[IPV6]: Reload skb->dst after xfrm6_route_forward()

While replacing dst_pmtu in ip6_output I found this little gem.  In
ip6_forward we're not reloading the dst pointer after calling
xfrm6_route_forward.  So all subsequent dereferences of dst will
refer to its pre-IPsec value.

The solution is of course to refresh its value.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:26:23-08:00 herbert
[IPSEC]: Fix room calculation in icmp_send

I'm now cleaning up all users of dst_pmtu with the aim of replacing
dst_pmtu with dst_mtu.  I'm going to start with the ones that actually
fix bugs.

This patch fixes the length calculation in icmp_send.  As it is
we're overestimating the space available by including the space
that would be used up by IPsec encapsulation.

IPv6 doesn't have this problem since its calculation is based
on 1280 instead of the PMTU.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:25:12-08:00 herbert
[IPSEC]: Handle local_df in IPv4

When cleaning up the remaining users of dst_pmtu I noticed that
local_df wasn't being treated correctly in IPsec.  In fact, if
you socket's dst went over IPsec, local_df is essentailly ignored.

This patch fixes that.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:22:31-08:00 bdschuym
[EBTABLES]: Fix smp race.

The patch below fixes an smp race that happens on such systems under
heavy load.
This bug was reported and solved by Steve Herrell
<steve_herrell@yahoo.ca>

Signed-off-by: Bart De Schuymer <bdschuym@pandora.be>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/14 21:21:24-08:00 tgraf
[IPV6]: Use dev_get_flags() while building inet6 ifinfo message

Use dev_get_flags() in inet6_fill_ifinfo() to fetch interface flags
to ensure correctly reporting IFF_PROMISC and IFF_ALLMULTI flags.

Signed-off-by: Thomas Graf <tgraf@suug.ch>
Signed-off-by: David S. Miller <davem@davemloft.net>

BKrev: 42368a6e1fCjhwp-WMHqUCadrfzYlQ