[IPV4]: Kill remaining unnecessary uses of dst_pmtu
Once again here is a couple of trivial dst_pmtu/dst_mtu replacements.
In both locations, we can only have simple dst entries which means
that dst == dst->path.

BTW, this is the rule that we should apply in future for uses of
dst_mtu/dst_pmtu (or other metrics on dst).  If the only dst's that
can appear are simple dst's (dst == dst->path), then we should use
dst_mtu or dst_metric.  If dst != dst->path, then whoever is writing
the code will need to think about which of dst or dst->path is the
right one.

In most instances dst will be the one.  However, as we have seen in
ip_append_data, dst->path may be needed rarely.  In particular, if
we're doing fragmentation immediately after IPsec, then you may need
it.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

BKrev: 423b2050UnSiWVVjD_Ixd1SSYIIW5A