[PATCH] ppc32: Fix FEC ethernet intialization on MPC8540 ADS board
The PHY interrupt for the DM9121 PHY connected to the FEC ethernet port
does not work on the MPC8540 ADS board.  If we tell the driver that the PHY
does not have an interrupt the FEC works properly on the MPC8540 ADS board.

Signed-off-by: Roy Zang
Signed-off-by: Kumar Gala <kumar.gala@freescale.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 423b49ec6HHx5i_MCSXVHuK2B_f3Ag