[PATCH] ppc64: fix kprobes calling smp_processor_id when preemptible
When booting with kprobes and preemption both enabled and
CONFIG_DEBUG_PREEMPT=y, I get lots of warnings about smp_processor_id being
called in preemptible code, from kprobe_exceptions_notify.  On ppc64,
interrupts and preemption are not disabled in the handlers for most
synchronous exceptions such as breakpoints and page faults (interrupts are
disabled in the very early exception entry code but are reenabled before
calling the C handler).

This patch adds a preempt_disable/enable pair to kprobe_exceptions_notify,
and moves the preempt_disable() in kprobe_handler() to be done only in the
case where we are about to single-step an instruction.  This eliminates the
bug warnings.

From: Ananth N Mavinakayanahalli <ananth@in.ibm.com>

The patch is fine, but it seems to break jprobes - we have an unbalanced
preempt_enable/disable path while handling jprobes. Patch below is
against 2.6.11-mm4 and fixes the issue.

Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Ananth N Mavinakayanahalli <ananth@in.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 423b4a8fnXzc8u6vC49Z1es4Z1ERTQ