[PATCH] ppc64: NUMA memory fixup (another try)
Below is a new version of the patch that allows holes within nodes on ppc64
NUMA.  I would appreciate it if someone familiar with OF device tree
parsing could take a look at this part of the code.  So far, I've gotten
this wrong twice.  Patch was tested in various configurations on a G5 and
OpenPower 720.

Signed-off-by: Mike Kravetz <kravetz@us.ibm.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 423b4b13rLfBmv951R8yASqetqRnZQ