Merge bk://kernel.bkbits.net/davem/net-2.6 into ppc970.osdl.org:/home/torvalds/v2.6/linux
2005/03/30 16:58:16-08:00 herbert
[NETLINK]: Fix sk_rmem_alloc assertion failure.

In netlink_dump we're operating on sk after dropping the cb lock.
This is racy because the owner of the socket could close it after
we drop the cb lock.

This is possible because netlink_dump isn't always called from the
context of the process that owns the socket.  For instance, if there
is contention on rtnl then rtnetlink requests will be processed by
the process that owns the rtnl.

The solution is to hold a ref count on the socket before we drop
the cb lock.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/30 16:39:35-08:00 davem
Merge bk://kernel.bkbits.net/tgraf/net-2.6-tcf_exts
into sunset.davemloft.net:/home/davem/src/BK/net-2.6

2005/03/30 16:35:41-08:00 yoshfuji
[AF_UNIX]: unix_mkname comment

Resurrect an old comment, explaining why this isn't an off-by-one error.

Signed-off-by: Hideaki YOSHIFUJI <yoshfuji@linux-ipv6.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/30 16:35:08-08:00 yoshfuji
[IPV6]: Fix address/interface handling according to the scoping architecture

I think this has been there for long time (maybe since 2.4...).

With the following patch, I can connect local link-local address.
- Change incoming interface according to the scoping architecture
- Choose source address on appropriate interface, according to the
  scoping architecture.

Signed-off-by: Hideaki YOSHIFUJI <yoshfuji@linux-ipv6.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/30 16:34:22-08:00 herbert
[PKT_SCHED]: Memory leak in ipt.c

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/30 16:33:25-08:00 rusty
[NETFILTER]: Restore ports module parameter for ip_nat_{ftp,irq}

There is no 'ports' parameter for the ip_nat_ftp and ip_nat_irc modules in
2.6.11: the ports parameter supplied to the
ip_conntrack_ftp/ip_conntrack_irc module defines the ports.  It was
unfortunate that we were lazy in the original implementation, and forced
the user to duplicate the arguments.

Even more unfortunate, the removal of the parameter caused autoloading to
break for various setups, with an 'Unknown parameter' message.  The
solution is to restore the parameter as a dummy, with a polite warning
message that it is no longer neccessary.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/30 16:32:22-08:00 jan.kiszka
[NET]: NULL pointer bug in netpoll.c

It seems that there is a gremlin sleeping in net/core/netpoll.c:find_skb().
 Even if no more buffers are available through skbs, skb is
dereferenced anyway.  The tiny patch should fix it.

Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/30 01:34:30+02:00 tgraf
[PKT_SCHED]: Fix action statistics dumping in compatibility mode

Extends the action dumping function by a parameter to differ between
regular calls and the one supposed to add the backward compatiblity
bits for old userspace applications.

Signed-off-by: Thomas Graf <tgraf@suug.ch>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/29 02:55:11+02:00 tgraf
[NET]: Make primary TLV type optional

Allows the use of the gnet_stats API for backward compatiblity
cases where no "modern" TLV structure is needed.

Signed-off-by: Thomas Graf <tgraf@suug.ch>
Signed-off-by: David S. Miller <davem@davemloft.net>

2005/03/29 02:45:52+02:00 tgraf
Cset exclude: hadi@cyberus.ca|ChangeSet|20050325173452|50562

BKrev: 424b6d0evYJNVK8ItG9ANN_pp67xQQ