[PATCH] ppc32: MPC8555 CPM2 size/pointers for FCCs aka "All-ones problem"
I've seen that MPC8555 support also made it into linux-2.6 by now.

I don't have hardware and time to test, but it looks like this bug has made
it over from the linuxppc-2.4 tree.  The 8541/8555 has a smaller CPM DPRAM
than the standard CPM2, just like the 8272.

Unfortunately, the manual recommends you to use a segment of CPM DPRAM that
is not implemented in the 8541/55.  If the smaller DPRAM is not taken into
account, fcc_enet.c will initialise the internal CPM buffer pointers
(fcc_riptr/fcc_tiptr) to invalid offsets, with the result that all the FCC
will ever send and receive are 0xff-en.

See also "All-ones problem with FCC1 on MPC8541" ~October 2004 on
linuxppc-embedded.

Signed-off-by: Stefan Nickl <Stefan.Nickl@kontron.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>

BKrev: 424c25daV7puLIkCm1RD6iaENhkTCQ